server/python-memory/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
cohere==5.3.0
numpy>=1.26.2
sqlalchemy==2.0.27
rank-bm25==0.2.2
nltk==3.8.1